fixed formatting problem
