+++ core/jvm/src/main/scala/zio/sql/Sql.scala
@@ -10,7 +10,8 @@ trait Sql
     with ExprModule
     with TableModule
     with InsertModule
-    with UtilsModule {
+    with UtilsModule
+    with InsertUtilsModule {
   self =>
 
   /*
